update login and reset password

+++ src/main/java/com/alister/dentistManager/service/impl/SingInServiceImpl.java
@@ -4,6 +4,7 @@ package com.alister.dentistManager.service.impl;
 import com.alister.dentistManager.dto.models.Doctor;
 import com.alister.dentistManager.repository.DoctorRepository;
 import com.alister.dentistManager.service.SingInService;
+import com.alister.dentistManager.web.exception.DoctorNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -28,16 +29,22 @@ public class SingInServiceImpl implements SingInService {
 
     @Override
     public Optional<Doctor> getDoctorById(Long id) {
-        return repository.findById(id);
+        return Optional.ofNullable(repository.findById(id).orElseThrow(() -> new DoctorNotFoundException(id)));
     }
 
     @Override
     public Doctor updateDoctor(Doctor doctor) {
+        if (!repository.existsById(doctor.getId())) {
+            throw new DoctorNotFoundException(doctor.getId());
+        }
         return repository.save(doctor);
     }
 
     @Override
     public void deleteDoctor(Long id) {
+        if (!repository.existsById(id)) {
+            throw new DoctorNotFoundException(id);
+        }
         repository.deleteById(id);
     }
 }
